Fix real positive cppcheck issues while maintaining VS2008 C++03 compatibility
- Fix printf format string mismatches with explicit casts:
  * CvTacticalAI.cpp:10645: Add (unsigned long) casts for size_t values in printf
  * CvDllDatabaseUtility.cpp:502: Add (int) cast for unsigned int in printf
- Fix opposite inner conditions by removing dead code:
  * CvPlayer.cpp: Remove 2 instances of unreachable code blocks
- Fix broken code structure in CvCultureClasses.cpp:
  * Remove extra closing braces that broke switch statement structure
  * Restore proper code block nesting for WLTKD tourism notifications

Total issues fixed: 8 out of 58 real positive cppcheck warnings
Maintains compatibility with Visual Studio 2008 SP1 and C++03/TR1 standards

+++ CvGameCoreDLL_Expansion2/CvDllDatabaseUtility.cpp
@@ -499,7 +499,7 @@ void CvDllDatabaseUtility::DatabaseRemapper()
 										bFirst = false;
 									}
 									char szUpdate[512];
-									sprintf_s(szUpdate, "UPDATE %s SET ID = %d WHERE ID = %d;", szTableName, ui, vTableIDs[ui]);
+									sprintf_s(szUpdate, "UPDATE %s SET ID = %d WHERE ID = %d;", szTableName, (int)ui, vTableIDs[ui]);
 									DB.Execute(szUpdate);
 								}
 							}
